requirements.txt
apache-airflow-providers-amazon==8.19.0
psycopg2-binary==2.9.9
boto3==1.34.0